+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.28.0](https://github.com/Substra/substra-tools/releases/tag/0.18.0) - 2022-09-26
+
 ### Added
 
 - feat: allow CLI parameters to be read from a file
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### CHANGED
 
 - BREAKING CHANGES:
+
   - the opener only exposes `get_data` and `fake_data` methods.
   - the results of the above method is passed under the `datasamples` keys within the `inputs` dict arg of all
     tools methods (train, predict, aggregate, score).